refactor(Profile): RHICOMPL-1782 split computed fields
